#720 RunningModuleをリファクタリングした
